Add "And Open" toggle to the Add Card dialog
This introduces another flow to the AddCardActivity dialog, which is
mutually exclusive with the "Multiples" option.  In this mode, after a
new card is added we then send an Intent to open the card in the main
Trello app (if installed) or else in the browser.  This allows for
subsequent rich editing of the card and adding things like description,
checkboxes, attachments, etc. without having to reimplement all that in
our dialog.

Signed-off-by: Jim Ramsay <i.am@jimramsay.com>